split query handler part of #524
