+++ lib/helpers/blacklist.js
@@ -2,25 +2,25 @@
 
 /*
 * Add a property clientProfile to true if the current property of the profile contains a client
-* @param {Object} profile - profile of the candidate
+* @param {Object} profileCan - profile of the candidate
 * @param {Array} clientList - list of clients
 */
 
-let blacklist = function(profile, clientList) {
+let blacklist = function(profileCandidate, clientList) {
 
-  profile.clientProfile = false;
+  profileCandidate.clientProfile = false;
   //lower case the list of companies
   let list = clientList.map(client => {return client.toLowerCase()} );
-  if(list.indexOf(profile.current.toLowerCase()) !== -1) {
-    profile.clientProfile = true;
+  if(list.indexOf(profileCandidate.current.toLowerCase()) !== -1) {
+    profileCandidate.clientProfile = true;
   } else {
     list.forEach(function(client) {
-      if(profile.current.toLowerCase().indexOf(client) !== -1) {
-        profile.clientProfile = true;
+      if(profileCandidate.current.toLowerCase().indexOf(client) !== -1) {
+        profileCandidate.clientProfile = true;
       }
     })
   }
-  return profile;
+  return profileCandidate;
 }
 
 module.exports = blacklist;
