+++ lib/helpers/delete_spaces.js
@@ -1,4 +1,7 @@
+'use strict';
+
 module.exports = function deleteSpaces (e) {
+
   if (e !== undefined) {
     return e.split(' ').filter( k => k !== '').join(' ');
   } else {
