+++ lib/helpers/match_terms.js
@@ -1,3 +1,11 @@
+'use strict';
+
+/**
+* Return number of results
+* @param {String} query e.g Js dev
+* @param {String} field - ES field e.g: fullname, headline, location, current
+*/
+
 module.exports = function(query, field) {
   var result = 0;
   var termsSearched = new Set();
@@ -21,5 +29,6 @@ module.exports = function(query, field) {
       termsSearched.add(term);
     }
   });
+  
   return result;
 }
